requirements.txt
# KalybrateX Dependencies

# Anthropic SDK for Claude API
anthropic>=0.18.0

# HTTP client
httpx>=0.27.0

# Browser automation for SkillsMP scraping
playwright>=1.40.0

# Data validation
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
